Rename the references in the reducers.
