+++ web/app/containers/home/reducer.js
@@ -2,7 +2,7 @@ import {handleActions} from 'redux-actions'
 import {fromJS} from 'immutable'
 import {mergePayload} from '../../utils/reducer-utils'
 
-import {receiveHomeData} from '../../integration-manager/responses'
+import {receiveHomeData} from '../../integration-manager/results'
 
 const CATEGORY_PLACEHOLDER_COUNT = 6
