+++ web/app/store/checkout/reducer.js
@@ -2,13 +2,13 @@ import Immutable from 'immutable'
 import {handleActions} from 'redux-actions'
 import {mergePayload} from '../../utils/reducer-utils'
 import {receiveCheckoutData, receiveShippingMethodInitialValues} from './actions'
-import * as integrationManagerResponses from '../../integration-manager/checkout/responses'
+import * as integrationManagerResults from '../../integration-manager/checkout/results'
 
 const productReducer = handleActions({
     [receiveCheckoutData]: mergePayload,
     [receiveShippingMethodInitialValues]: mergePayload,
-    [integrationManagerResponses.receiveShippingMethodInitialValues]: mergePayload,
-    [integrationManagerResponses.receiveCheckoutData]: mergePayload
+    [integrationManagerResults.receiveShippingMethodInitialValues]: mergePayload,
+    [integrationManagerResults.receiveCheckoutData]: mergePayload
 }, Immutable.Map())
 
 export default productReducer
